refactor: centralize icons, fix type errors, and improve timed editions UI
Migrate all Lucide icon imports to a centralized src/lib/icons.ts module
with a reusable Icon component. Fix type errors across the codebase for
clean builds. Add date-time picker and calendar UI components for timed
editions. Improve post detail page layout, BuyButton wallet reconnect
flow, and edition options form UX. Clean up server function boundaries
and update API route auth patterns.

+++ src/server/services/blockchain/addressUtils.ts
@@ -104,7 +104,7 @@ export function addressToBytes(addr: string): Uint8Array {
 		throw new Error(`Address encoding produced ${bytes.length} bytes, expected 32`);
 	}
 
-	return bytes;
+	return bytes as Uint8Array;
 }
 
 /**
